ml_api/requirements.txt
Flask==2.0.1
Tensorflow==2.6.0
Pillow==8.3.2
